requirements.txt
torch==1.11.0
pyyaml==6.0
pillow==9.1.1
torchvision==0.12.0
omegaconf==2.2.1
